Extract playback logic to hook
